FIx y-tree, fix value drift for both trees

+++ src/main/java/dev/kalarianathecila/schoolStuff/tutelini/Turtle.java
@@ -66,7 +66,7 @@ class Turtle
 
   }
 
-  public void back(int length){
+  public void back(double length){
       turnLeft(180);
       stopDraw();
       move(length);
